dodajanje receptov

+++ recipes.php
@@ -19,7 +19,7 @@ include_once 'db.php';
         echo '<a href="recipe.php?id='.$row['id'].'">';
             echo $row['recipe'].' ('.$row['category'].')';
         echo '</a>';
-        echo ' <a href="recipe_delete.php?id='.$row['id'].'">(x)</a>';
+        echo ' <a href="recipe_delete.php?id='.$row['id'].'" onclick="return confirm(\'Prepričani?\')">(x)</a>';
         echo ' <a href="recipe_edit.php?id='.$row['id'].'">(u)</a>';
         echo '</li>';
     }
